+++ compat-driver-5.x/src/test/java/com/github/starnowski/jamolingo/compat/driver/operators/expand/nestedtree/ExpandOperatorWithHandlingTreeRelationsTest.java
@@ -520,6 +520,87 @@ public class ExpandOperatorWithHandlingTreeRelationsTest extends AbstractItTest 
                                                     }]
                                                     """,
             JSONCompareMode.STRICT_ORDER),
+        // Expand level handle by $lookup stage and use the $apply
+        Arguments.of(
+            TREETYPE1_MONGO_COLLECTION_USAGE_INFO,
+            Set.of(1),
+            "$expand=treeType2s($apply=concat(filter(categoryId eq 2),groupby((categoryId))))",
+            """
+                                                            [{ "_id": 1, "index": 1, "parentId": null, "categoryId": 1,
+                                                             "treeType2s": [
+                                                             { "_id": 3, "index": 3, "parentId": 2, "categoryId": 2, "treeType1Id": 1 },
+                                                             { "categoryId": 1},
+                                                             { "categoryId": 2}
+                                                             ]
+                                                            }]
+                                                            """,
+            JSONCompareMode.LENIENT),
+        // Expand level handle by $lookup stage and use the $apply with order so that the
+        // NON_EXTENSIBLE mode could be used for comparison
+        Arguments.of(
+            TREETYPE1_MONGO_COLLECTION_USAGE_INFO,
+            Set.of(1),
+            "$expand=treeType2s($apply=concat(filter(categoryId eq 2),groupby((categoryId)));$orderby=categoryId asc,index desc)",
+            """
+                            [{ "_id": 1, "index": 1, "parentId": null, "categoryId": 1,
+                            "treeType2s": [
+                                        { "categoryId": 1},
+                                        { "_id": 3, "index": 3, "parentId": 2, "categoryId": 2, "treeType1Id": 1 },
+                                        { "categoryId": 2}
+                                    ]
+                            }]
+         """,
+            JSONCompareMode.NON_EXTENSIBLE),
+        // Expand level handle by $lookup stage and use the $apply with order so that the
+        // NON_EXTENSIBLE mode could be used for comparison
+        Arguments.of(
+            TREETYPE1_MONGO_COLLECTION_USAGE_INFO,
+            Set.of(1),
+            "$expand=treeType2s($apply=concat(filter(categoryId eq 2),groupby((categoryId),aggregate($count as totalCount)));$orderby=categoryId asc,index desc)",
+            """
+                                    [{ "_id": 1, "index": 1, "parentId": null, "categoryId": 1,
+                                    "treeType2s": [
+                                                { "categoryId": 1, "totalCount": 2},
+                                                { "_id": 3, "index": 3, "parentId": 2, "categoryId": 2, "treeType1Id": 1 },
+                                                { "categoryId": 2, "totalCount": 1}
+                                            ]
+                                    }]
+                 """,
+            JSONCompareMode.NON_EXTENSIBLE),
+        // Expand level handle by $lookup stage and use the $apply with order so that the
+        // NON_EXTENSIBLE mode could be used for comparison
+        // Filtering also the dynamic field like totalCount
+        Arguments.of(
+            TREETYPE1_MONGO_COLLECTION_USAGE_INFO,
+            Set.of(1),
+            "$expand=treeType2s($apply=groupby((categoryId),aggregate($count as totalCount))/filter(totalCount eq 1))",
+            """
+                                            [{ "_id": 1, "index": 1, "parentId": null, "categoryId": 1,
+                                            "treeType2s": [
+                                                        { "categoryId": 2, "totalCount": 1}
+                                                    ]
+                                            }]
+                         """,
+            JSONCompareMode.NON_EXTENSIBLE),
+        // $expand=treeType2s($apply=groupby((categoryId),aggregate($count as
+        // totalCount));$filter=totalCount eq 1)
+        // Olingo does not supports filtering by dynamic properties
+        // https://issues.apache.org/jira/browse/OLINGO-1303
+        //            Arguments.of(
+        //                    TREETYPE1_MONGO_COLLECTION_USAGE_INFO,
+        //                    Set.of(1),
+        //                    "$expand=treeType2s($apply=groupby((categoryId),aggregate($count as
+        // totalCount))/filter(totalCount eq 1))",
+        //                    """
+        //                                                    [{ "_id": 1, "index": 1, "parentId":
+        // null, "categoryId": 1,
+        //                                                    "treeType2s": [
+        //                                                                { "categoryId": 2,
+        // "totalCount": 1}
+        //                                                            ]
+        //                                                    }]
+        //                                 """,
+        //                    JSONCompareMode.NON_EXTENSIBLE),
         Arguments.of(
             TREETYPE1_MONGO_COLLECTION_USAGE_INFO,
             Set.of(10),